requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
freezegun==1.5.1
gunicorn==20.0.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
